feat: edit data to Firestore
fixed file structure too
