+++ nearbnb/src/components/logicism/Dot.js
@@ -1,0 +1,11 @@
+export default function Dot({ color }) {
+  const style = {
+    height: 25,
+    width: 25,
+    margin: "0px 10px",
+    backgroundColor: color,
+    borderRadius: "50%",
+    display: "inline-block"
+  };
+  return <span style={style}></span>;
+}
